# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/posts:
    get:
      tags:
        - Posts
      summary: List Posts
      description: Get a paginated list of Posts with optional filtering and sorting
      operationId: listPosts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Posts
      summary: Create Post
      description: Create a new Post
      operationId: createPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/post_tags:
    get:
      tags:
        - PostTags
      summary: List PostTags
      description: Get a paginated list of PostTags with optional filtering and sorting
      operationId: listPostTags
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostTagList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PostTags
      summary: Create PostTag
      description: Create a new PostTag
      operationId: createPostTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostTagRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostTag'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Post:
      type: object
      required:
        - id
        - blog_id
        - website_id
        - title
        - slug
        - is_published
        - post_date
        - visits
        - allow_comments
        - metadata
      properties:
        id:
          type: string
          format: uuid
        blog_id:
          type: string
          format: uuid
        website_id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        slug:
          type: string
          maxLength: 120
        content:
          nullable: true
          type: string
        teaser:
          nullable: true
          type: string
          maxLength: 500
        cover:
          nullable: true
          type: object
        author_officer:
          nullable: true
          type: string
          format: uuid
        author_name:
          nullable: true
          type: string
          maxLength: 120
        is_published:
          type: boolean
        published_date:
          nullable: true
          type: string
          format: date-time
        post_date:
          type: string
          format: date-time
        visits:
          type: integer
          format: int32
        allow_comments:
          type: boolean
        archived_at:
          nullable: true
          type: string
          format: date-time
        archived_by_blog_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreatePostRequest:
      type: object
      required:
        - blog_id
        - website_id
        - title
        - slug
        - is_published
        - post_date
        - visits
        - allow_comments
        - metadata
      properties:
        blog_id:
          type: string
          format: uuid
        website_id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        slug:
          type: string
          maxLength: 120
        content:
          nullable: true
          type: string
        teaser:
          nullable: true
          type: string
          maxLength: 500
        cover:
          nullable: true
          type: object
        author_officer:
          nullable: true
          type: string
          format: uuid
        author_name:
          nullable: true
          type: string
          maxLength: 120
        is_published:
          type: boolean
        published_date:
          nullable: true
          type: string
          format: date-time
        post_date:
          type: string
          format: date-time
        visits:
          type: integer
          format: int32
        allow_comments:
          type: boolean
        archived_at:
          nullable: true
          type: string
          format: date-time
        archived_by_blog_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    PostList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PostTag:
      type: object
      required:
        - id
        - post_id
        - tag_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        post_id:
          type: string
          format: uuid
        tag_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreatePostTagRequest:
      type: object
      required:
        - post_id
        - tag_id
        - metadata
      properties:
        post_id:
          type: string
          format: uuid
        tag_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    PostTagList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PostTag'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []